Revert login hook
